// src/types/breadcrumb.enum.ts
export enum BreadcrumbTitle {
  Datasource = "Datasource",
  Create = "Create",
  Register = "Register",
  Application = "Application",
  DetailsApplication = "Details Application",
  IntegrationsApplication = "Integrations Application",
  SecurityApplication = "Security Application",
  CreateApplication = "Create Application",
  CreateDatasource = "Create Datasource",
  DetailsDatasource = "Details Datasource",
  IntegrationsDatasource = "Integrations",
  SecurityDatasource = "Security",
  Users = "Users",
  UsersInRole = "Users in role",
  UsersCreate = "Create User",
  Attributes = "Attributes",
  Roles = "Roles",
  Groups = "Groups",
  Actions = "Actions",
  Details = "Details",
  RolesMapping = "Roles Mapping",
  Credentials = "Credentials",
  RolesCreate = "Create Role",
  Settings = "Settings",
  Mail = "Email",
  Sessions = "Sessions",

  // Titles สำหรับ Billing
  TemplateConfiguration = "Template Configuration",
  ReconcileConfiguration = "Reconcile Configuration",
  ProcessTemplate = "Process Template",
  ProcessReconcile = "Process Reconcile",
  Master = "Master",
  ParameterConfig = "Parameter Configuration Master",
  HolidaysCalendar = "Holidays Calendar Master",
  BillingSchedule = "Billing Schedule Master",
  GroupBill = "Group Bill Master",
  Process = "Process",
  CustomerRegistration = "Customer Registration",
  IncomeLedgerManagement = "Income Ledger Management",
  CustomerInfoManagement = "Customer Info Management",
  RequestToBilling = "Request to Billing",
  BillingPeriodManagement = "Billing Period Management",
  InvoiceManagement = "Invoice Management",
  InvoicePreparationPrinting = "Invoice Preparation and Printing",
  ClosePeriod = "Close Period",
  GenerateSpecialInvoice = "Generate Special Invoice",
  Report = "Report",
  BillingTransactionReport = "Billing Transaction Report",
  BillingTransactionReport2 = "Billing Transaction Report 2",
  AccountTransactionReport = "Account Transaction Report",
  CountCustomerCircuitReport = "Count Customer and Circuit Report",
  DeferredIncomeReport = "Deferred Income Report",
  AccruedIncomeReport = "Accrued Income Report",
  BillingTransactionCompareReport = "Billing Transaction Compare Report",
  AgingReport = "Aging report",
  RevenueSharingERP = "Revenue Sharing from ERP",
  PaymentTransactionERP = "Payment Transaction from ERP",
}