// src/types/path.enum.ts
export enum Path {
  Home = "/",
  Album = "/album",
  Docs = "/docs.html",
  Console = "/console",
  Login = "/login",
  ForgotPassword = "/forgot-password",
  CheckYourEmail = "/check-your-email",
  SecurityVerification = "/security-verification",
  ResetTwoFactor = "/reset-two-factor",
  Datasource = "/console/datasource",
  DatasourceDetail = "/console/datasource/details",
  DatasourceCreate = "/console/datasource/create",
  Application = "/console/application",
  CreateApplication = "/console/application/create",
  DetailApplication = "/console/application/details",
  IntegrationApplication = "/console/application/integrations",
  SecurityApplication = "/console/application/security",
  ApplicationRegister = "/console/application/register",
  Users = "/console/users",
  UsersDetails = "/console/users/details",
  UsersCreate = "/console/users/create",
  UsersAttributes = "/console/users/attributes",
  UsersCredentials = "/console/users/credentials",
  UsersActions = "/console/users/actions",
  UsersGroups = "/console/users/groups",
  UsersRoleMapping = "/console/users/role-mapping",
  UsersSessions = "/console/users/sessions",
  Roles = "/console/roles",
  Groups = "/console/groups",
  Actions = "/console/actions",
  RolesDetails = "/console/roles/details",
  RolesCreate = "/console/roles/create",
  RolesAttributes = "/console/roles/attributes",
  RolesGroups = "/console/roles/groups",
  RolesUsers = "/console/roles/users-in-role",
  Forbidden = "/forbidden",
  Unauthorized = "/unauthorized",
  Settings = "/console/settings",
  Email = "/console/settings/email",

   // Paths 
   TemplateConfiguration = "/template-configuration",
   ReconcileConfiguration = "/reconcile-configuration",
   ProcessTemplate = "/process-template",
   ProcessReconcile = "/process-reconcile",

}